// veroDigit - Frontend/src/app/Model/Artwork.ts
export interface Artwork {
  id: number;
  title: string;
  main_reference_number: string;
  department_title: string;
  artist_title: string;
  artist_display: string;
  place_of_origin: string;
  medium_display: string;
  technique_titles: string[];
  term_titles: string[];
  date_start: string;
  image_id: string;
}